move dilation2d to image
