+++ topi/python/topi/generic/default.py
@@ -14,28 +14,21 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
+# pylint: disable=invalid-name,unused-argument
+"""The default schedule used by various operators"""
+import tvm
+from tvm import te
 
-# pylint: disable=wildcard-import
-"""Generic declaration and schedules.
 
-This is a recommended way of using TOPI API.
-To use the generic schedule function, user must set
-the current target scope using with block. See also :any:`tvm.target`
-
-Example
--------
-.. code-block:: python
-
-  # create schedule that dispatches to topi.cuda.schedule_injective
-  with tvm.target.create("cuda"):
-    s = tvm.tir.generic.schedule_injective(outs)
-"""
-from __future__ import absolute_import as _abs
-
-from .nn import *
-from .injective import *
-from .extern import *
-from .vision import *
-from .sort import *
-from .search import *
-from .image import *
+def default_schedule(outs, auto_inline):
+    """Default schedule for llvm."""
+    target = tvm.target.Target.current(allow_none=False)
+    outs = [outs] if isinstance(outs, te.tensor.Tensor) else outs
+    if target.target_name != "llvm":
+        raise RuntimeError("schedule not registered for '%s'" % target)
+    s = te.create_schedule([x.op for x in outs])
+    if auto_inline:
+        x = outs[0]
+        te.schedule.AutoInlineInjective(s)
+        s[x].fuse(s[x].op.axis)
+    return s
