+++ topi/python/topi/generic/search.py
@@ -17,7 +17,8 @@
 # pylint: disable=invalid-name, no-member
 """Generic search operators"""
 from __future__ import absolute_import as _abs
-from .vision import _default_schedule
+from .default import default_schedule as _default_schedule
+
 
 def schedule_argwhere(outs):
     """Schedule for argwhere operator.
