+++ topi/python/topi/image/__init__.py
@@ -20,3 +20,4 @@
 from __future__ import absolute_import as _abs
 
 from .resize import *
+from .dilation2d import *
